Refactor: Introduce _internal package for private interface
Moves all non-public modules and sub-packages of `aipproto` into an `aipproto._internal` package to clearly delineate the public vs. private API.

Changes include:
- Created `src/aipproto/_internal/` directory.
- Moved `hierarchy.py`, `method_group.py`, `options.py`, `render.py`, `methods/`, `req_res/`, and `templates/` into the `_internal` package.
- Updated all import statements across the codebase to reflect the new locations, using absolute imports (e.g., `from aipproto._internal import ...`).
- Updated `importlib.resources.files` path in `render.py` to point to `aipproto._internal.templates`.
- Added `__all__` to `src/aipproto/__init__.py` to explicitly define the exported public symbols.
