+++ src/aipproto/__init__.py
@@ -1,3 +1,11 @@
+__all__ = [
+    "generate_file_content",
+    "Namespace",
+    "ResourceConfig",
+    "UpdateConfig",
+    "DeleteConfig",
+]
+
 from aipproto.generate import generate_file_content
 from aipproto.resource import Namespace
 from aipproto.resource_config import ResourceConfig, UpdateConfig, DeleteConfig
