+++ src/aipproto/_internal/__init__.py
@@ -1,0 +1,2 @@
+# This file makes Python treat the `_internal` directory as a package.
+# It can be left empty.
