+++ src/aipproto/_internal/hierarchy.py
