+++ src/aipproto/_internal/methods/__init__.py
@@ -1,0 +1,2 @@
+# This file makes Python treat the `_internal/methods` directory as a package.
+# It can be left empty.
