+++ src/aipproto/_internal/req_res/update_req_res.py
@@ -1,13 +1,14 @@
 from typing import List
-from aipproto import options, render, resource
+from aipproto import resource
+from aipproto._internal import options, render
 
 
 def from_resource(resource_type: resource.Resource) -> List[render.ReqRes]:
     pascal = resource_type.format_type("pascal")
     return [
         render.ReqRes(
-            type=f"Delete{pascal}Request",
-            description=f"Request message for deleting a {pascal}.",
+            type=f"Update{pascal}Request",
+            description=f"Request message for updating a {pascal}.",
             fields=_fields(resource_type),
         ),
     ]
@@ -17,26 +18,23 @@ def _fields(resource_type: resource.Resource) -> List[render.ReqResField]:
     pascal = resource_type.format_type("pascal")
     fields = [
         render.ReqResField(
-            type="string",
-            name="name",
+            type=resource_type.format_type("pascal"),
+            name=resource_type.format_type("snake"),
             comment_lines=[
-                f"The name of the {pascal} to delete.",
+                f"The {pascal} being updated.",
             ],
             options=[
                 options.field_behavior("REQUIRED"),
-                options.resource_reference(
-                    "type", f"{resource_type.namespace().name}/{pascal}"
-                ),
             ],
         )
     ]
-    if resource_type.has_children:
+    if resource_type.config().update_config().partial:
         fields.append(
             render.ReqResField(
-                type="bool",
-                name="force",
+                type="google.protobuf.FieldMask",
+                name="update_mask",
                 comment_lines=[
-                    f"If set, deletes the {pascal} and all of its children.",
+                    "The set of fields to update.",
                 ],
                 options=[
                     options.field_behavior("OPTIONAL"),
