+++ src/aipproto/_internal/templates/.gitkeep
@@ -1,0 +1,2 @@
+# This file ensures Git tracks the templates directory.
+# It can be removed after files are added if desired.
