+++ src/aipproto/_internal/templates/proto.jinja
